[omxplayer] Reduce rate of clock change to reduce video stutters
After analysing some files with video stutter, it was found that the filter coefficients used
for adjusting the clock were set higher than the maximum hdmi clock deviation allowed.

This means when gpu makes small clock adjustments to correct for lip-sync, the video drops frames.

With this patch the frame drops are fixed.