Fix audio getting stuck on seeking
The bug was actually caused by m_iBufferOutputUsed (OMXAudioCodecOMX.cpp) not being set back to 0 on flush and therefore the old values remain in m_dts and m_pts on the next Decode (OMXAudioCodecOMX.cpp). They are then written back into dts and pts in GetData(OMXAudioCodecOMX.cpp) and voila the timestamps are wrong.