[omxplayer] Don't sync up to passthrough audio packets - let GPU handle it
This code was just broken.
When sync is lost it attempts to identify how many bytes to skip to regain sync,
but does nothing with that skip count - just drops the whole frame.

It also stops EAC3 from being passed through.

The GPU has similar sync detection code which does avoid dropping the whole frame,
so just pass the packets through and let GPU deal with it.