Pass SubmitEOS message through decode queue
So, the problem is that SubmitEOS is called when demuxer reaches end of file, but not all the encoded frames have been submitted to decoder.

m_packets may contain dozens of encoded frames of data at the point SubmitEOS is called. We want those to all be submitted to decoder before the SubmitEOS is called.